web_app/requirements_web.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2
numpy>=1.24
torch>=2.0
timm>=0.9
albumentations>=1.3
opencv-python-headless>=4.8
Pillow>=10
pandas>=2.0
h5py>=3.9
scikit-learn>=1.3
tqdm>=4.65
